Update 08_using_compose.md
Hi so,
with the current state of the tutorial I kept getting the error 
"service 'volumes' must be a mapping not an array". after doing some search I got to the docker compose section of the docker site and landed on this finding 

volumes:
  logvolume01: {}

So I tweaked it to this and ran docker-compose up -d and it ran fine no errors. 
Thanks for letting me contribute :)
You guys are awesome!